ucs-api/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
protobuf==3.20.1
pyformance==0.4
pyparsing==3.0.7
PyYAML==6.0
requests==2.28.1
signalfx==1.1.15
six==1.16.0
sseclient-py==1.7.2
ucsmsdk==0.9.12
urllib3==1.26.10
ws4py==0.5.1
